refactor: Move DataService health/version availability logic into ConfigurationRepository
- Moved the deprecated healthAccess, checkVersion, and isPlanetAvailable code paths from DataService.kt to ConfigurationRepository.kt and ConfigurationRepositoryImpl.kt.
- Moved the checkCheckSum method from DataService.kt to ConfigurationRepository.kt and ConfigurationRepositoryImpl.kt.
- Added new EntryPoint files (`ConfigurationEntryPoint.kt`, `UserEntryPoint.kt`) to provide dependencies to non-Hilt classes.
- Updated call sites in AutoSyncWorker.kt, BaseResourceFragment.kt, SettingActivity.kt, SyncActivity.kt, MainApplication.kt, BecomeMemberActivity.kt, DialogUtils.kt, and LoginSyncManager.kt to use the appropriate repository or injected DataService.
- Removed the deprecated and moved methods from DataService.
- Removed manual instantiation of DataService and relied on Hilt injection.
- Deleted the now-unused RepositoryEntryPoint.kt.
