+++ app/src/main/java/org/ole/planet/myplanet/di/UserEntryPoint.kt
@@ -1,0 +1,12 @@
+package org.ole.planet.myplanet.di
+
+import dagger.hilt.EntryPoint
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import org.ole.planet.myplanet.repository.UserRepository
+
+@EntryPoint
+@InstallIn(SingletonComponent::class)
+interface UserEntryPoint {
+    fun userRepository(): UserRepository
+}
