+++ app/src/main/java/org/ole/planet/myplanet/repository/ConfigurationRepository.kt
@@ -8,6 +8,7 @@ interface ConfigurationRepository {
     fun checkHealth(listener: SuccessListener)
     fun checkVersion(callback: CheckVersionCallback, settings: SharedPreferences)
     fun checkServerAvailability(callback: PlanetAvailableListener?)
+    suspend fun checkCheckSum(path: String?): Boolean
 
     interface CheckVersionCallback {
         fun onUpdateAvailable(info: MyPlanet?, cancelable: Boolean)
