+++ app/src/main/java/org/ole/planet/myplanet/repository/ConfigurationRepositoryImpl.kt
@@ -28,8 +28,10 @@ import org.ole.planet.myplanet.utilities.VersionUtils
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-
+import java.io.IOException
 import org.ole.planet.myplanet.di.AppPreferences
+import org.ole.planet.myplanet.utilities.FileUtils
+import org.ole.planet.myplanet.utilities.Sha256Utils
 
 class ConfigurationRepositoryImpl @Inject constructor(
     @ApplicationContext private val context: Context,
@@ -86,6 +88,26 @@ class ConfigurationRepositoryImpl @Inject constructor(
         } catch (e: Exception) {
             e.printStackTrace()
             listener.onSuccess("Health access initialization failed")
+        }
+    }
+
+    override suspend fun checkCheckSum(path: String?): Boolean = withContext(Dispatchers.IO) {
+        try {
+            val response = apiInterface.getChecksum(UrlUtils.getChecksumUrl(preferences)).execute()
+            if (response.isSuccessful) {
+                val checksum = response.body()?.string()
+                if (!checksum.isNullOrEmpty()) {
+                    val f = FileUtils.getSDPathFromUrl(context, path)
+                    if (f.exists()) {
+                        val sha256 = Sha256Utils().getCheckSumFromFile(f)
+                        return@withContext checksum.contains(sha256)
+                    }
+                }
+            }
+            false
+        } catch (e: IOException) {
+            e.printStackTrace()
+            false
         }
     }
