+++ app/src/main/java/org/ole/planet/myplanet/service/AutoSyncWorker.kt
@@ -69,7 +69,7 @@ class AutoSyncWorker(
     }
 
     override fun onUpdateAvailable(info: MyPlanet?, cancelable: Boolean) {
-        startDownloadUpdate(context, UrlUtils.getApkUpdateUrl(info?.localapkpath), null, workerScope)
+        startDownloadUpdate(context, UrlUtils.getApkUpdateUrl(info?.localapkpath), null, workerScope, configurationRepository)
     }
 
     override fun onCheckingVersion() {}
